Added logic to statusReadout
This commit adds a few new logic checks as well as new global variables to be used between the doLaunchTasks function and the updateDynamicReadouts function, in order to make so that the readout will tell the user when ignition, liftoff, and staging is occurring. Includes a slight change to make so that target pitch is specified to be in degrees.