Fixed Issue #12
Fixed the doReleaseTasks function so that it actually does put the throttle at 0 at the end of the long. This makes so that if the Override Throttle button was used during the ascent, it will still be set to the end so that the user gets an accurate insertion to the apoapsis they selected, instead of having the engine continue burning until the fuel runs out, if they left the throttle above zero. The if-statement checks to see if the Abort Launch button was used, and it doesn't do anything with the pilot main throttle if it was. This way if you decided to abort the launch but want to keep controlling manually, you won't have to get the throttle set back to where you had it. It only goes to zero if you never used the Override Throttle button.